Split Nav into two files
I decided to completely change the stylistic direction of the navbar. On <360px screens it will remain how it is, but on wider screens it will start beneathe the splash and stick to the top of the screen once scrolled past. Anticipating this change, I have seperated this navbar into its own separate script within a navbar component folder.