Contact Footer
Contact Section has been reduced to a footer bar with my media links. I think this makes a better UX becuase signals the end of the app, no more to scroll. And if someone wants to contact me, they probably arent waiting for a paragraph beckoning them to do so-- they'll just hit the links. Also made a lot of responsive-size changes with both splash and contact media links.